+++ tests/test_tipc/configs/llama/train_infer_python.txt
@@ -13,7 +13,7 @@ null:null
 null:null
 ##
 trainer:norm_train
-norm_train:../examples/language_model/llama/finetune_instruction_generation.py --model_name_or_path facebook/llama-7b-2l --do_train --max_steps 500 --recompute False --benchmark --overwrite_output_dir --output_dir ./checkpoints/ --fp16_opt_level O2 --learning_rate 3e-5 --lr_scheduler_type constant --warmup_steps 0 --seed 23 --logging_steps 1 --max_grad_norm -1
+norm_train:../examples/language_model/llama/finetune_generation.py --model_name_or_path facebook/llama-7b-2l --do_train --max_steps 500 --recompute False --benchmark --overwrite_output_dir --output_dir ./checkpoints/ --fp16_opt_level O2 --learning_rate 3e-5 --lr_scheduler_type constant --warmup_steps 0 --seed 23 --logging_steps 1 --max_grad_norm -1
 pact_train:null
 fpgm_train:null
 distill_train:null
